Move `ImageCommentCellController` to separate file
